#1023 advanced data components
